requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
amqp==5.2.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
boto3==1.33.10
botocore==1.33.13
cachetools==5.3.2
celery==5.3.6
celery-types==0.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.0
cryptography==42.0.7
dataclasses-json==0.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
distro==1.9.0
dj-database-url==2.1.0
Django==5.0
django-ace==1.32.0
django-filter==23.5
django-import-export==3.3.3
django-object-actions==4.2.0
django-split-settings==1.2.0
django-storages==1.14.2
django-stubs==4.2.7
django-stubs-ext==5.0.2
djangorestframework==3.14.0
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
docx2txt==0.8
drf-spectacular==0.26.5
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
flupy==1.2.0
frozenlist==1.4.1
fsspec==2024.5.0
gigachat==0.1.9
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.64.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.3.7
langchain==0.2.1
langchain-community==0.2.4
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-qdrant==0.1.0
langchain-text-splitters==0.2.0
langsmith==0.1.67
llama-index==0.9.13
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
odfpy==1.4.1
openai==1.31.0
openpyxl==3.1.3
orjson==3.10.3
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
pgvector==0.1.8
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
portalocker==2.8.2
prompt_toolkit==3.0.45
protobuf==5.27.1
psycopg==3.1.14
psycopg-c==3.1.14
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.18.0
PyJWT==2.8.0
pypdf==3.17.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
qdrant-client==1.9.1
redis==5.0.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
tablib==3.5.0
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
vecs==0.4.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.9.4
